+++ _receiver/C/rtl_rec.h
@@ -5,7 +5,7 @@ struct rtl_rec_struct {
     int dev_ind, gain;
     rtlsdr_dev_t *dev;
     uint8_t *buffer;   
-    unsigned long long buff_ind;
+    unsigned long buff_ind;
     pthread_t async_read_thread;        
     uint32_t center_freq, sample_rate;
 };
